+++ src/Providers/AdvancedApiControllerServiceProvider.php
@@ -4,10 +4,9 @@ namespace LaravelAdvancedApiController\Providers;
 
 use Illuminate\Support\ServiceProvider;
 use LaravelAdvancedApiController\Http\Controllers\APIController;
-use LaravelAdvancedApiController\Console\Commands\MakeAdvancedPolicy;
-use LaravelAdvancedApiController\Console\Commands\MakeAdvancedController;
-use LaravelAdvancedApiController\Console\Commands\MakePolicy;
 use LaravelAdvancedApiController\Console\Commands\MakeController;
+use LaravelAdvancedApiController\Console\Commands\MakeModel;
+use LaravelAdvancedApiController\Console\Commands\MakePolicy;
 
 class AdvancedApiControllerServiceProvider extends ServiceProvider
 {
@@ -32,10 +31,9 @@ class AdvancedApiControllerServiceProvider extends ServiceProvider
 
 			// Enregistrement des commandes personnalisées
 			$this->commands([
-				MakeAdvancedPolicy::class,
-				MakeAdvancedController::class,
-				MakePolicy::class,
 				MakeController::class,
+				MakeModel::class,
+				MakePolicy::class,
 			]);
 		}
 	}